addressing varyvol's suggestions
